feat: Implement NativeFridaBridge for macOS ARM64
- Added NativeFridaBridge class for macOS ARM64 platform, implementing FridaBridge interface.
- Integrated Frida RPC management with asynchronous handling of messages and responses.
- Introduced methods for environment preparation, gadget injection, and JDWP injection.
- Enhanced error handling and logging for Frida operations.
- Refactored existing Unix NativeFridaBridge code to separate RPC management and message handling.
- Created OnFridaMessage.kt for handling Frida messages and parsing JSON responses.
